requirements.txt
Django==1.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-guardian==1.2.4
django-ordered-model==0.4.0
djangorestframework==3.0.1
python-social-auth==0.2.1
django-mailgun==0.2.2
django-filter==0.9.1

